+++ src/main/java/com/finalProject/Back/service/CafeService.java
@@ -5,6 +5,7 @@ import com.finalProject.Back.dto.response.Cafe.RespCafeDto;
 import com.finalProject.Back.entity.Cafe.Cafe;
 import com.finalProject.Back.entity.Cafe.CafeDetail;
 import com.finalProject.Back.entity.Cafe.CafeLike;
+import com.finalProject.Back.entity.User;
 import com.finalProject.Back.entity.board.Board;
 import com.finalProject.Back.repository.CafeLikeMapper;
 import com.finalProject.Back.repository.CafeMapper;
@@ -26,9 +27,10 @@ public class CafeService {
     @Autowired
     private CafeLikeMapper cafeLikeMapper;
 
-    public int addCafe(ReqCafeDto dto) {
+    public Long addCafe(ReqCafeDto dto) {
         System.out.println("카페 정보" + dto);
-        int cafeId = cafeMapper.save(dto.toEntity());
+        Long cafeId = cafeMapper.save(dto.toEntity());
+        // cafe 추가성공 건수를 누가넣어놨냐?
         return cafeMapper.saveDefaultCafeMenu(cafeId);
     }
